+++ middleman-core/lib/middleman-core/preview_server/server_information.rb
@@ -69,13 +69,11 @@ module Middleman
         @port         = config[:port]
         @server_name  = config[:server_name]
         @https        = config[:https]
-        @web_server   = config[:web_server]
 
         config[:bind_address] = bind_address
         config[:port]         = port
         config[:server_name]  = server_name
         config[:https]        = https?
-        config[:web_server]   = web_server
       end
 
       # Make information of internal server class avaible to make debugging
@@ -151,11 +149,6 @@ module Middleman
       # Is https enabled?
       def https?
         @https == true
-      end
-
-      # For https works only webrick
-      def web_server
-        https? ? :webrick : @web_server&.to_sym
       end
     end
   end
